join team bug solved
